requirements.base.txt
fastapi
uvicorn
python-multipart
python-dotenv
pillow
numpy
opencv-python-headless